+++ nitinol-eventsource/src/process/tell_target.rs
@@ -1,7 +1,7 @@
 use futures_core::future::BoxFuture;
+use nitinol_contract::Aggregate;
 use nitinol_persistence::AggregateId;
 
-use crate::aggregate::Aggregate;
 use crate::decider::Decider;
 use crate::error::TellError;
 use crate::process::proxy::AggregateProxy;